requirements.txt
streamlit
google-generativeai
opencv-python
pytesseract
pdfplumber
Pillow
